requirements.txt
python-binance
#boto3
#TA-Lib
